requirements.txt
streamlit
pandas
google-generativeai
faiss-cpu
torch
torchvision
torchaudio
sentence-transformers==2.2.2
transformers==4.36.2
huggingface-hub==0.19.4